% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterResults.R
\name{filterResults}
\alias{filterResults}
\title{Filter results on specific classification level}
\usage{
filterResults(
  results,
  classification_level = c("Category", "Main class", "Sub class", "Species",
    "Molecular subspecies", "Structural subspecies", "Isomeric subspecies")
)
}
\arguments{
\item{results}{`data.frame` Result data.frame from SwissLipidsR}

\item{classification_level}{`character` Level of classification that shall be returned}
}
\value{
`data.frame` A filtered data.frame is returned
}
\description{
`filterResults` performs filtering of results obtained from other functions
}
\examples{

results <- swissLipidsAdvancedSearch(formula = "C39H76NO8P")
filterResults(results, classification_level = "Species")
}
\author{
Michael Witting
}
