% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{swissLipidsAdvancedSearch}
\alias{swissLipidsAdvancedSearch}
\title{SwissLipids API advanced search function}
\usage{
swissLipidsAdvancedSearch(
  name = NA_character_,
  smiles = NA_character_,
  inchikey = NA_character_,
  formula = NA_character_,
  mz = NA_real_,
  adduct = NA_character_,
  massErrorRate = NA_real_
)
}
\arguments{
\item{name}{`character`}

\item{smiles}{`character`}

\item{inchikey}{`character`}

\item{formula}{`character`}

\item{mz}{`numeric`}

\item{adduct}{`character`}

\item{massErrorRate}{`numeric`}
}
\value{
`data.frame` A named list with the search results
}
\description{
`swissLipidsAdvancedSearch` Performs an advanced search using the SwissLipids API
}
\examples{

swissLipidsAdvancedSearch(name = "PC(34:2)")
swissLipidsAdvancedSearch(formula = "C39H76NO8P")

swissLipidsAdvancedSearch(mz = 410.243, adduct = "MassExact", massErrorRate = 0.001)
}
\author{
Michael Witting#'
}
