% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{swissLipidsGetChildren}
\alias{swissLipidsGetChildren}
\title{SwissLipids API basic search function}
\usage{
swissLipidsGetChildren(entity_id)
}
\arguments{
\item{entity_id}{`character` Search term for which children entries shall  be
searched in SwissLipids}
}
\description{
`swissLipidsGetChildren` Performs a search for potential children
}
\examples{

swissLipidsGetChildren("SLM:000000352")
}
\author{
Michael Witting#'
}
