% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{swissLipidsMapping}
\alias{swissLipidsMapping}
\title{SwissLipids API Mapping}
\usage{
swissLipidsMapping(
  from = c("SwissLipids", "LipidMaps", "ChEBI", "HMDB"),
  to = c("SwissLipids", "LipidMaps", "ChEBI", "HMDB", "UniProtKB", "Rhea"),
  ids
)
}
\arguments{
\item{from}{`character` Abbrevation of data from which the ids shall be mapped}

\item{to}{`character` Abbrevation of database to which the ids shall be mapped}

\item{ids}{`character` A vector of IDs that shall be mapped.}
}
\value{
`list` A named list of the same length as ids is returned.
}
\description{
`swissLipidsMapping` uses the SwissLipids API to perform mapping between
    different database identifier
}
\examples{

ids <- c("SLM:000048885", "SLM:000000651")
from <- "SwissLipids"
to <- "LipidMaps"

swissLipidsMapping(from, to, ids)
}
\author{
Michael Witting
}
