% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{swissLipidsSearch}
\alias{swissLipidsSearch}
\title{SwissLipids API basic search function}
\usage{
swissLipidsSearch(term, type = c("metabolite", "protein"))
}
\arguments{
\item{term}{`character` Search term which shall be searched in SwissLipids}

\item{type}{`character` Type of search term, either metabolite or protein}
}
\value{
`data.frame` A data.frame with the search results
}
\description{
`swissLipidsSearch` Performs a basic search using the SwissLipids API
}
\examples{

swissLipidsSearch("Phosphatidate (36:2)")
}
\author{
Michael Witting#'
}
